requirements.txt
PySide6==6.10.0
PySide6_Addons==6.10.0
PySide6_Essentials==6.10.0
shiboken6==6.10.0
python-escpos
pyusb
python-barcode
Pillow